+++ packages/react-ui-smoke-test/cra-template-react-ui/template/src/App.tsx
@@ -46,7 +46,6 @@ import {
   TooltipMenu,
 } from '@skbkontur/react-ui';
 import EditIcon from '@skbkontur/react-icons/Edit';
-import { FLAT_THEME } from '@skbkontur/react-ui/lib/theming/themes/FlatTheme';
 
 export const App = () => {
   // TODO App must be pure render function
@@ -65,7 +64,7 @@ export const App = () => {
       <Checkbox checked={false} onValueChange={() => ({})}>
         Checkbox
       </Checkbox>
-      <ComboBox getItems={q => Promise.resolve([{ value: 1, label: 'First' }])} />
+      <ComboBox getItems={(q) => Promise.resolve([{ value: 1, label: 'First' }])} />
       <CurrencyLabel value={12356.1} currencySymbol={'₽'} />
       <DateInput />
       <DatePicker value={''} onValueChange={() => ({})} />
@@ -132,7 +131,7 @@ export const App = () => {
       </SidePage>
       <Spinner type="big" caption="big" />
       <Sticky side="top" getStop={() => null}>
-        {_ => 'Small loan of a million dollars'}
+        {(_) => 'Small loan of a million dollars'}
       </Sticky>
       <Switcher items={['One', 'Two', 'Three']} />
       <Tabs value="fuji">
@@ -142,7 +141,7 @@ export const App = () => {
       <Textarea />
       <Toggle />
       <Token>Default</Token>
-      <TokenInput type={TokenInputType.Combined} getItems={_ => Promise.resolve(['First'])} />
+      <TokenInput type={TokenInputType.Combined} getItems={(_) => Promise.resolve(['First'])} />
       <Tooltip render={() => <div />} pos="right top">
         a
       </Tooltip>
@@ -150,11 +149,8 @@ export const App = () => {
         <MenuHeader>Заголовок меню</MenuHeader>
       </TooltipMenu>
       <LocaleContext.Provider value={{ langCode: LangCodes.ru_RU }}>
-        <LocaleContext.Consumer>{locale => locale.langCode}</LocaleContext.Consumer>
+        <LocaleContext.Consumer>{(locale) => locale.langCode}</LocaleContext.Consumer>
       </LocaleContext.Provider>
-      <ThemeContext.Provider value={ThemeFactory.create(FLAT_THEME)}>
-        <ThemeContext.Consumer>{theme => theme.btnFontSizeMedium}</ThemeContext.Consumer>
-      </ThemeContext.Provider>
     </main>
   );
 };