+++ packages/react-ui/.creevey/config.js
@@ -1,4 +1,6 @@
 const path = require('path');
+const axios = require('axios');
+require('dotenv').config({ path: '../../.env' });
 
 /**
  * Debuggin instructions: https://wiki.skbkontur.ru/pages/viewpage.action?pageId=418699157
@@ -14,12 +16,18 @@ const capabilities = debug
     }
   : {};
 
+const resolverStorybookUrl = (port) => ({
+  storybookUrl: `http://localhost:${port}`,
+  resolveStorybookUrl:
+    process.env.GET_IP_URL && (() => axios(process.env.GET_IP_URL).then((res) => `http://${res.data}:${port}`)),
+});
+
 const config = {
+  ...resolverStorybookUrl(6060),
   storybookDir: path.join(__dirname, '../.storybook'),
   reportDir: path.join(__dirname, 'report'),
   screenDir: path.join(__dirname, 'images'),
-  gridUrl: 'https://frontinfra:frontinfra@grid.testkontur.ru/wd/hub',
-  storybookUrl: 'http://localhost:6060',
+  gridUrl: process.env.GRID_URL,
   // NOTE Should refactor Button styles without 1px-border
   maxRetries: process.env.TEAMCITY_VERSION ? 10 : 0,
   babelOptions: (options) => ({
@@ -28,34 +36,14 @@ const config = {
   }),
   diffOptions: { threshold: 0, includeAA: false },
   browsers: {
-    chrome: {
-      browserName: 'chrome',
-      viewport: { width: 1024, height: 720 },
-      platformName: 'linux',
-      _storybookGlobals: {
-        theme: 'DEFAULT_THEME_OLD',
-      },
-      name: 'infrafront/chrome',
-      ...capabilities,
-    },
     chrome8px: {
       browserName: 'chrome',
       viewport: { width: 1024, height: 720 },
       platformName: 'linux',
       _storybookGlobals: {
-        theme: 'DEFAULT_THEME',
+        theme: 'DEFAULT_THEME_8PX_OLD',
       },
       name: 'infrafront/chrome8px',
-      ...capabilities,
-    },
-    chromeFlat: {
-      browserName: 'chrome',
-      viewport: { width: 1024, height: 720 },
-      platformName: 'linux',
-      _storybookGlobals: {
-        theme: 'FLAT_THEME_OLD',
-      },
-      name: 'infrafront/chromeFlat',
       ...capabilities,
     },
     chromeFlat8px: {
@@ -63,91 +51,82 @@ const config = {
       viewport: { width: 1024, height: 720 },
       platformName: 'linux',
       _storybookGlobals: {
-        theme: 'FLAT_THEME',
+        theme: 'FLAT_THEME_8PX_OLD',
       },
       name: 'infrafront/chromeFlat8px',
       ...capabilities,
     },
-    firefox: {
-      browserName: 'firefox',
+    chrome: {
+      browserName: 'chrome',
       viewport: { width: 1024, height: 720 },
+      platformName: 'linux',
       _storybookGlobals: {
-        theme: 'DEFAULT_THEME_OLD',
+        theme: 'DEFAULT_THEME',
       },
-      name: 'infrafront/firefox',
+      name: 'infrafront/chrome',
       ...capabilities,
     },
     firefox8px: {
       browserName: 'firefox',
       viewport: { width: 1024, height: 720 },
       _storybookGlobals: {
-        theme: 'DEFAULT_THEME',
+        theme: 'DEFAULT_THEME_8PX_OLD',
       },
       name: 'infrafront/firefox8px',
-      ...capabilities,
-    },
-    firefoxFlat: {
-      browserName: 'firefox',
-      viewport: { width: 1024, height: 720 },
-      _storybookGlobals: {
-        theme: 'FLAT_THEME_OLD',
-      },
-      name: 'infrafront/firefoxFlat',
       ...capabilities,
     },
     firefoxFlat8px: {
       browserName: 'firefox',
       viewport: { width: 1024, height: 720 },
       _storybookGlobals: {
-        theme: 'FLAT_THEME',
+        theme: 'FLAT_THEME_8PX_OLD',
       },
       name: 'infrafront/firefoxFlat8px',
       ...capabilities,
     },
-    ie11: {
-      browserName: 'internet explorer',
-      viewport: { width: 1024, height: 720 },
-      _storybookGlobals: {
-        theme: 'DEFAULT_THEME_OLD',
-      },
-      name: 'infrafront/ie11',
-      ...capabilities,
-
-      // NOTE Enable after switch new separate pool for IE to allow test hover
-      // 'se:ieOptions': {
-      //   enablePersistentHover: true,
-      //   nativeEvents: true,
-      //   requireWindowFocus: true,
-      //   'ie.usePerProcessProxy': true,
-      //   'ie.browserCommandLineSwitches': '-private',
-      //   'ie.ensureCleanSession': true,
-      // },
-    },
-    ie118px: {
-      browserName: 'internet explorer',
+    firefox: {
+      browserName: 'firefox',
       viewport: { width: 1024, height: 720 },
       _storybookGlobals: {
         theme: 'DEFAULT_THEME',
       },
-      name: 'infrafront/ie118px',
+      name: 'infrafront/firefox',
       ...capabilities,
     },
-    ie11Flat: {
+    // NOTE Enable after switch new separate pool for IE to allow test hover
+    // 'se:ieOptions': {
+    //   enablePersistentHover: true,
+    //   nativeEvents: true,
+    //   requireWindowFocus: true,
+    //   'ie.usePerProcessProxy': true,
+    //   'ie.browserCommandLineSwitches': '-private',
+    //   'ie.ensureCleanSession': true,
+    // },
+    ie118px: {
       browserName: 'internet explorer',
       viewport: { width: 1024, height: 720 },
       _storybookGlobals: {
-        theme: 'FLAT_THEME_OLD',
+        theme: 'DEFAULT_THEME_8PX_OLD',
       },
-      name: 'infrafront/ie11Flat',
+      name: 'infrafront/ie118px',
       ...capabilities,
     },
     ie11Flat8px: {
       browserName: 'internet explorer',
       viewport: { width: 1024, height: 720 },
       _storybookGlobals: {
-        theme: 'FLAT_THEME',
+        theme: 'FLAT_THEME_8PX_OLD',
       },
       name: 'infrafront/ie11Flat8px',
+      ...capabilities,
+    },
+    ie11: {
+      browserName: 'internet explorer',
+      viewport: { width: 1024, height: 720 },
+      _storybookGlobals: {
+        theme: 'DEFAULT_THEME',
+      },
+      name: 'infrafront/ie11',
       ...capabilities,
     },
   },