+++ packages/react-ui/components/ComboBox/__stories__/Combobox.stories.tsx
@@ -14,6 +14,8 @@ import { Gapped } from '../../Gapped';
 import { MenuHeader } from '../../MenuHeader';
 import { delay } from '../../../lib/utils';
 import { Tooltip } from '../../Tooltip';
+import { ThemeContext } from '../../../lib/theming/ThemeContext';
+import { ThemeFactory } from '../../../lib/theming/ThemeFactory';
 
 const { getCities } = require('../__mocks__/getCities.js');
 
@@ -28,7 +30,12 @@ SimpleComboboxStory.storyName = 'simple combobox';
 
 SimpleComboboxStory.parameters = {
   creevey: {
-    skip: [{ in: ['ie11', 'ie11Flat', 'ie118px', 'ie11Flat8px'], tests: ['hovered', 'selected_2', 'select_1'] }],
+    skip: [
+      {
+        in: ['ie11', 'ie118px', 'ie11Flat8px'],
+        tests: ['hovered', 'selected_2', 'select_1'],
+      },
+    ],
     tests: {
       async plain() {
         await this.expect(await this.takeScreenshot()).to.matchImage('plain');
@@ -272,7 +279,7 @@ OpenToTop.storyName = 'open to top';
 
 OpenToTop.parameters = {
   creevey: {
-    skip: [{ in: ['ie11', 'ie11Flat', 'ie118px', 'ie11Flat8px'], tests: 'hovered' }],
+    skip: [{ in: ['ie11', 'ie118px', 'ie11Flat8px'], tests: 'hovered' }],
     tests: {
       async plain() {
         const element = await this.browser.findElement({ css: '[data-tid="container"]' });
@@ -1125,4 +1132,44 @@ WithTooltip.parameters = {
       },
     },
   },
+};
+
+export const MobileSimple = () => (
+  <ThemeContext.Consumer>
+    {(theme) => {
+      return (
+        <ThemeContext.Provider
+          value={ThemeFactory.create(
+            {
+              mobileMediaQuery: '(max-width: 576px)',
+            },
+            theme,
+          )}
+        >
+          <SimpleCombobox />
+          <div style={{ height: 15 }} />
+          <ComplexCombobox />
+          <div style={{ height: 15 }} />
+          <TestComboBox
+            onSearch={search}
+            renderItem={renderValue}
+            renderAddButton={(query) =>
+              query && (
+                <MenuItem key={'mobileAddButton'} isMobile onClick={() => alert(query)}>
+                  Добавить {query}
+                </MenuItem>
+              )
+            }
+          />
+        </ThemeContext.Provider>
+      );
+    }}
+  </ThemeContext.Consumer>
+);
+MobileSimple.title = 'Mobile combobox stories';
+MobileSimple.parameters = {
+  viewport: {
+    defaultViewport: 'iphone',
+  },
+  creevey: { skip: [true] },
 };