+++ packages/react-ui/components/Switcher/Switcher.styles.ts
@@ -18,7 +18,7 @@ export const styles = memoizeStyle({
     `;
   },
 
-  label(t: Theme) {
+  caption(t: Theme) {
     return css`
       color: ${t.switcherTextColor};
       vertical-align: middle;
@@ -26,27 +26,27 @@ export const styles = memoizeStyle({
     `;
   },
 
-  labelSmall(t: Theme) {
+  captionSmall(t: Theme) {
     return css`
-      margin-right: ${t.switcherLabelGapSmall};
-      font-size: ${t.switcherLabelFontSizeSmall};
-      line-height: ${t.switcherLabelLineHeightSmall};
+      margin-right: ${t.switcherCaptionGapSmall};
+      font-size: ${t.switcherCaptionFontSizeSmall};
+      line-height: ${t.switcherCaptionLineHeightSmall};
     `;
   },
 
-  labelMedium(t: Theme) {
+  captionMedium(t: Theme) {
     return css`
-      margin-right: ${t.switcherLabelGapMedium};
-      font-size: ${t.switcherLabelFontSizeMedium};
-      line-height: ${t.switcherLabelLineHeightMedium};
+      margin-right: ${t.switcherCaptionGapMedium};
+      font-size: ${t.switcherCaptionFontSizeMedium};
+      line-height: ${t.switcherCaptionLineHeightMedium};
     `;
   },
 
-  labelLarge(t: Theme) {
+  captionLarge(t: Theme) {
     return css`
-      margin-right: ${t.switcherLabelGapLarge};
-      font-size: ${t.switcherLabelFontSizeLarge};
-      line-height: ${t.switcherLabelLineHeightLarge};
+      margin-right: ${t.switcherCaptionGapLarge};
+      font-size: ${t.switcherCaptionFontSizeLarge};
+      line-height: ${t.switcherCaptionLineHeightLarge};
     `;
   },