+++ packages/react-ui/components/Switcher/Switcher.tsx
@@ -25,7 +25,7 @@ export interface SwitcherProps extends CommonProps {
 
   onValueChange?: (value: string) => void;
 
-  label?: string;
+  caption?: string;
 
   error?: boolean;
 
@@ -59,7 +59,7 @@ export class Switcher extends React.Component<SwitcherProps, SwitcherState> {
         }),
       ),
     ]).isRequired,
-    label: PropTypes.string,
+    caption: PropTypes.string,
     value: PropTypes.string,
     onValueChange: PropTypes.func,
   };
@@ -94,12 +94,12 @@ export class Switcher extends React.Component<SwitcherProps, SwitcherState> {
       className: styles.input(),
     };
 
-    const lableClassName = cx(styles.label(this.theme), this.getLabelSizeClassName());
+    const captionClassName = cx(styles.caption(this.theme), this.getLabelSizeClassName());
 
     return (
       <CommonWrapper {...this.props}>
         <div>
-          {this.props.label ? <div className={lableClassName}>{this.props.label}</div> : null}
+          {this.props.caption ? <div className={captionClassName}>{this.props.caption}</div> : null}
           <div className={styles.wrap()}>
             <input {...inputProps} />
             <div className={listClassName}>
@@ -210,12 +210,12 @@ export class Switcher extends React.Component<SwitcherProps, SwitcherState> {
   private getLabelSizeClassName = (): string => {
     switch (this.props.size) {
       case 'large':
-        return styles.labelLarge(this.theme);
+        return styles.captionLarge(this.theme);
       case 'medium':
-        return styles.labelMedium(this.theme);
+        return styles.captionMedium(this.theme);
       case 'small':
       default:
-        return styles.labelSmall(this.theme);
+        return styles.captionSmall(this.theme);
     }
   };
 }